Add chat domain repository and interactor
