+++ app/src/main/java/uddug/com/naukoteka/mvvm/chat/ChatCreateSingleViewModel.kt
@@ -10,14 +10,14 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
-import uddug.com.data.repositories.chat.ChatRepository
 import uddug.com.domain.entities.profile.UserProfileFullInfo
+import uddug.com.domain.interactors.chat.ChatInteractor
 import uddug.com.domain.repositories.user_profile.UserProfileRepository
 import javax.inject.Inject
 
 @HiltViewModel
 class ChatCreateSingleViewModel @Inject constructor(
-    private val chatRepository: ChatRepository,
+    private val chatInteractor: ChatInteractor,
     private val userProfileRepository: UserProfileRepository,
 ) : ViewModel() {
 
@@ -47,7 +47,7 @@ class ChatCreateSingleViewModel @Inject constructor(
     fun onUserClick(userId: Long) {
         viewModelScope.launch {
             try {
-                val dialogId = chatRepository.createDialog(userId)
+                val dialogId = chatInteractor.createDialog(userId)
                 _events.emit(ChatCreateSingleEvent.DialogCreated(dialogId))
             } catch (e: Exception) {
                 _uiState.value = ChatCreateSingleUiState.Error(e.message ?: "Unknown error")
