+++ app/src/main/java/uddug/com/naukoteka/ui/chat/di/ChatModule.kt
@@ -5,11 +5,11 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import retrofit2.Retrofit
-import uddug.com.data.NaukotekaCookieJar
 import uddug.com.data.cache.cookies.CookiesCache
-import uddug.com.data.repositories.chat.ChatRepository
 import uddug.com.data.repositories.chat.ChatRepositoryImpl
 import uddug.com.data.services.chat.ChatApiService
+import uddug.com.domain.interactors.chat.ChatInteractor
+import uddug.com.domain.repositories.chat.ChatRepository
 
 @Module
 @InstallIn(ViewModelComponent::class)
@@ -22,9 +22,10 @@ object ChatModule {
 
 
     @Provides
-    fun provideChatRepository(apiService: ChatApiService): ChatRepository {
-        return ChatRepositoryImpl(apiService)
-    }
+    fun provideChatRepository(apiService: ChatApiService): ChatRepository = ChatRepositoryImpl(apiService)
+
+    @Provides
+    fun provideChatInteractor(chatRepository: ChatRepository): ChatInteractor = ChatInteractor(chatRepository)
 
     @Provides
     fun provideSocketService(cookiesCache: CookiesCache): SocketService {
