+++ data/src/main/java/uddug/com/data/repositories/chat/ChatRepositoryImpl.kt
@@ -1,41 +1,15 @@
 package uddug.com.data.repositories.chat
 
-import io.reactivex.Single
 import uddug.com.data.mapper.mapChatDtoToDomain
-
 import uddug.com.data.services.chat.ChatApiService
-import uddug.com.data.services.models.response.chat.DialogInfoDto
 import uddug.com.data.services.models.response.chat.mapDialogInfoDtoToDomain
 import uddug.com.domain.entities.chat.Chat
 import uddug.com.domain.entities.chat.DialogInfo
 import uddug.com.domain.entities.chat.MediaMessage
 import uddug.com.domain.entities.chat.MessageChat
 import uddug.com.domain.entities.chat.updateOwnerInfoFromDialog
-import uddug.com.domain.entities.feed.PostComment
+import uddug.com.domain.repositories.chat.ChatRepository
 import javax.inject.Inject
-
-interface ChatRepository {
-    suspend fun getChats(): List<Chat>
-    suspend fun getMessages(
-        currentUserId: String,
-        dialogId: Long,
-        limit: Int = 50,
-    ): List<MessageChat>
-
-    suspend fun getDialogInfo(dialogId: Long): DialogInfo
-    suspend fun getMessagesWithOwnerInfo(
-        currentUserId: String,
-        dialogId: Long,
-        limit: Int = 50,
-    ): List<MessageChat>
-
-    suspend fun getDialogMedia(
-        dialogId: Long,
-    ): List<MediaMessage>
-
-    suspend fun createDialog(userId: Long): Long
-}
-
 
 class ChatRepositoryImpl @Inject constructor(
     private val apiService: ChatApiService,
